+++ CHANGELOG.md
@@ -8,6 +8,9 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Removed
 - Deprecated UartServer class.
 
+### Fixed
+- uart_echo_server example and test server not stopping the server task before destruction.
+
 ## [1.3.1] - 2026-08-18
 ### Fixed
 - RX FIFO interrupt configuration.